amun: processor/robotfilter: fix different signedness warning

+++ src/amun/processor/tracking/robotfilter.cpp
@@ -383,7 +383,7 @@ void RobotFilter::get(world::Robot *robot, const FieldTransform &transform, bool
         addPrimaryDetection(m_measurements.first());
     } else if (
         auto it = std::find_if(m_measurements.cbegin(), m_measurements.cend(), [this](const world::TransformedRobotMeasurement &p) {
-            return p.camera_id() == m_primaryCamera;
+            return (qint32)p.camera_id() == m_primaryCamera;
         });
         it != m_measurements.cend()
     ) {
